requirements.txt
ultralytics
fastapi
uvicorn[standard]
opencv-python
pillow
numpy
onnxruntime
torch
torchvision
torchaudio